Create EnemySpawner Game Object and Plan out GamePlay loop
Created an Enemy Spawner gameObject and script that will spawn the enemies based on the rounds of the game. Enemy Spawner should deal with all instances of spawning and even removing all enemies. The GameManager will keep track of the status of the game and make calls to the functions of Enemy Spawner when necessary. Plan to remove all instances of spawning enemies in original GameManager.